requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
et_xmlfile==2.0.0
flatbuffers==25.1.24
fonttools==4.55.8
gast==0.4.0
google-auth==2.38.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.3
grpcio==1.70.0
h5py==3.12.1
idna==3.10
intel-extension-for-tensorflow==0.0.0.dev1
joblib==1.4.2
keras==3.8.0
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.8
numpy==2.0.2
oauthlib==3.2.2
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plot-model==0.20
protobuf==4.25.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydot==3.0.4
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
seaborn==0.13.2
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow_intel==2.18.0
tensorflow_keras==0.1
termcolor==2.5.0
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.5.0
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.1.3
wget==3.2
wrapt==1.17.2
